Related: cool#11942 desktop, vcl: add some logging on high priority tasks
The original problem around tile rendering, a large document with many
pages and a shape is resolved, but we got a report that the old bad
behavior is still observed from time to time, with no steps to
reproduce.

So add some logging to obtain more info: if the problem is that the
first tile gets rendered because of an unwanted high priority VCL task, then
let's log all "interrupting" tasks and also the doc-load end and the
tile-render start to see what tasks interrupt. Possibly that will point
out a new task that doesn't happen for me locally and then that can be
investigated.

Change-Id: I4c92bf7d5b42628398e586276f80052e0de46306
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189999
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190056
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>