Reset hyperlink dialog pages after link data loaded
Data about the selected link is loaded in the line GetBindings().Update(SID_HYPERLINK_GETLINK), which requires the page to already exist. This means we end up creating the 4 pages without the correct data to load into them. For selectively hiding fields based on the selection, it can be helpful to know the size of the other tabs, and therefore which fields in them are showing.

Therefore, we iterate through the tabs after the data is loaded and reset all of them to ensure each has the correct elements hidden.

Change-Id: I6a6a696455c6ff1980fe7d9926a24262b80c5ca7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190006
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190680
Tested-by: Jenkins