tdf#67797 autofmt: don't delete leading blanks without bSetNumRule
If autoFormat is not going to turn the text into a list,
  [m_aFlags.bSetNumRule for [T]]
or a style
  [! m_aFlags.bAFormatAsInput which is [M]]
then don't automatically delete the leading/following blanks.

This specifically became very inconsistent with
    A. List item 1
    B. List item 2
    C. List item 3   #DeleteLeadingTrailingBlanks called
since A and B are not recognized as numbering,
but C is considered part of a Roman numbering list.

But even
    1. List item #DeleteLeadingTrailingBlanks called
should not actually strip out the leading blanks
unless it is actually going to be changed into a list
(or if the style is going to be changed).

Note that "Delete spaces and tabs at beginning and end of paragraph"
  [m_aFlags.bAFormatByInpDelSpacesAtSttEnd for [T]]
is only supposed to apply if "Apply Styles" is also enabled.
  [m_aFlags.bReplaceStyles for [T]]

Change-Id: Ie381c65a8f664393dd3428b5615502c17fc9f228
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190043
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>