Automatically map generic JSON and XML documents to Calc
This includes one bug fix where orcus_xml did not call
import_factory's finalize() interface method per interface
contract.  With it properly fixed in liborcus itself, it
is no longer necessary to manually call it at the end of
ScOrcusXMLContextImpl::importXML().

Also moving up the type detection order for the generic
HTML type to be before the generic XML type since something
like

<html><body>some text</body></html>

can be detected as a valid XML if it is checked first.
A (well-formed) HTML can be considered a speciazlied XML,
so it should come before XML.

Change-Id: I20f24706421104181ad3dc8250d93f78d6b611e9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190042
Tested-by: Jenkins
Reviewed-by: Kohei Yoshida <kohei@libreoffice.org>