Throw from OText{Input,Output}Stream::setEncoding on unknown encoding
Before this, it was possible to pass an arbitrary string to the method,
and have no way of knowing about the problem. IllegalArgumentException
is derived from RuntimeException, which is explicitly documented to be
allowed at every interface method, signalling errors not covered by the
interface method specification, so this is not an API change.

Change-Id: I038f7765ec3f332f0e787351fa032feeef2cefb0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190048
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins