tdf#166724 - Checkbox anchor wrongly imported
Form Controls are not being "fully" imported as the MoveWithCell and SizeWithCell properties are not set internally in Calc.
Contrary to their own documentation, Microsoft Excel seems to declare (in the VML file) those properties with no value when they are false, and does not declare them when they are true.
Test xlsx documents generated with Excel 2007, 2010, 2013 and 365 have led me to this conclusion.

Change-Id: I7ba64983ae8036df8d672d6c94b2ca54d740a064
Signed-off-by: Johann LORBER <jlorber@linagora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190166
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Code-Style: Justin Luth <jluth@mail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194707
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>