tdf#127675 Treat Biff12 when pasting
Prior to Biff8, Biff12 is tested as format for pasting.

The new format is read using SfxObjectShell::DoLoad. Since the import
must be to an ScDocument in SCDOCMODE_CLIP mode, a new ctor was added
for SfxObjectShell, that initializes the document with needed mode.

The change in MS_Excel_2007_Binary.xcu enables GetFilter4ClipBoardId
to return the correct filter.

Since clipboard documents don't have a chart listener collection, we
must not call SetChartListenerCollectionNeedsUpdate from SetChanged;
otherwise, ScDocument::UpdateChartListenerCollection would crash.

Change-Id: If57b1ffb9a2f8d0573cf77b9eb82f81aa4250b5e
Co-authored-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190077
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>