tdf#104827 doc export: Keep CR when justified text before section
This patch only benefits Micrsoft Word.
MS Word has a bug where even the last line of the paragraph
is justified when it ends with a continuous break.

This change could probably be made
for (almost) all continuous section breaks,
but it likely has other implications or side effects,
so just minimize the change to this one specific use-case.

Change-Id: Ifd16b78a95d8cb52d37f948a6f6fe5fcdaf2a254
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190193
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>