Related: tdf#168073 Extract ScEditFieldObj insertion into a function
This makes it easier to see that ScCellObj::insertTextContent attempts
one of two alternative insertion methods. Additionally, this undoes a
change of mine, made in commit 1e040b72222012f9a5f1e47d19aaf0c0369e08ad
(tdf#168073: check if xRange is this, 2025-08-23), which disallowed to
insert the same ScEditFieldObj secont time. I did that to follow what
writer does; but the method doesn't specify that behavior, so it was an
unwarranted change.

Change-Id: Iea0d1ff5039f6e36126a93b0d56cd049ff070945
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190196
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>