tdf#168131: don't pick hybrid PDF, when normal PDF was detected
When a standard PDF is dragged onto a spreadsheet, a matcher is
created in ScViewFunc::PasteFile, with current document service.
Inside GuessFilterControlDefaultUI call, queryTypeByDescriptor
detects pdf_Portable_Document_Format type, and draw_pdf_import
as filter. Detection in PDFDetector::detect checks for embedded
hybrid stream, and if found, it would return one of "addstream"
filter names; and when we get a simple "*_pdf_import", it means
that hybrid PDF was not detected.

But SfxFilterMatcher::GuessFilterControlDefaultUI tried to find
a filter for the detected file type, that matches the document
service. It listed all filters for the type, and found the only
filter that matched Calc: calc_pdf_addstream_import, which is for
hybrid PDFs.

This change makes an explicit check, that when original detection
was for a normal PDF, we don't pick hybrid PDF filter.

Change-Id: Ibab8c9ccaf97637eebbf9f93b8618d9d5f2fe73d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190274
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins