tdf#168139 Avoid assertion failure on X11 with Skia/Vulkan
Using X11 backend (gen), LibreOffice produces assertion failure on a
debug build when running with Skia/Vulkan. That is because of missing
implementation for handling transparency on that platform.

I turned the assert/TODO into a warning to avoid assertion failure.
This is supposed to be only for X11, and is tested on Linux.

Change-Id: Iea7ac732123446843aaa71748c10bc17e305fc8f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190291
Tested-by: Jenkins
Reviewed-by: Hossein   <hossein@libreoffice.org>