tdf#167760 tdf#162326 limit reset para list attrs on style apply
Problem: applying any para style
was removing all numbering or bullets.

This fixes a 25.2 regression from
commit 075560420a7aa238182e2d80dfe1c5fbaad3edbd
  tdf#162326 reset paragraph list attributes on style apply

The problem is that bResetListAttrs was already set to true,
but the conditions under which bResetListAttrs applies
were changed by the above commit
so now it affects situations it was never intended to match.

Specifically, this was the condition that changed:
-    && pFormat->GetItemState(RES_PARATR_NUMRULE) == SfxItemState::SET
+    && pCNd->GetTextNode()->IsInList()

So two changes need to happen here:
1.) bResetListAttrs needs to only be set
under the existing comment's stated condition (docst.cxx/doc.hxx)
that the para style itself contains a numbering rule.

2.) When the control key is held down,
then bRestListAttrs needs to be set to true
(which is the stated intention of the above commit).

make CppunitTest_sw_uiwriter11 \
    CPPUNIT_TEST_NAME=testTdf167760_numberedPara

Change-Id: Iae646124a8a434dd3b49a75caeae4bb360396a49
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190309
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190321
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190488
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>