sw markdown export: handle <code>
The import side of this was added in commit
f6411560817dd9e2c07564eef100dcb8f995804f (tdf#162153 Add missing inline
code span, 2025-08-27), which maps MD_SPAN_CODE to the font of the
RES_POOLCOLL_HTML_PRE style (some kind of monospace font).

So do the opposite on export: if the font is the same as the monospace
font from RES_POOLCOLL_HTML_PRE, then export that portion of text as
`text`.

To help nesting, export the start of this as the last char prop, and
write the end of this as the first char prop.

Change-Id: I7ba21502a1040b08c6b66929ceedfd1fe72f9f70
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190325
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins