tdf#130857 qt weld: Implement TreeView tooltip handling
Handle the QEvent::ToolTip event for the
QTreeView's viewport. Retrieve the tooltip
for the item using weld::TreeView::signal_query_tooltip
and display it if it's non-empty.

This makes the file path of the corresponding template
show up when hovering over an item in the
"File" -> "Templates" -> "Manage Templates" dialog
with the list view mode enabled with
SAL_VCL_QT_USE_WELDED_WIDGETS=1, in a WIP branch
where support for that dialog using native Qt widgets
is declared.

Change-Id: Ia849a9fa6506ffbd2d74be00600e0a694f1b9efd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190317
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>