tdf#130857 qt weld: Pass parent when setting tree entry text on insert
Take the parent into account when creating the TreeIter
passed to QtInstanceTreeView::set_text.

This makes the "Tools" -> "AutoText" show the tree view
entries as expected again in a WIP branch where support for
using native Qt widgets with SAL_VCL_QT_USE_WELDED_WIDGETS=1
is declared.

This broke previously with

    commit 44e08e5b7fef8d9aa65e81e6a9a1367b177f473c
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sun Aug 3 22:36:37 2025 +0200

        tdf#130857 qt weld: Use text col in TreeView::insert

which had resulted in top-level entries always getting used
instead of the children of the given parent.

Change-Id: I0b63ccba426e737c6bae968cd9dc8c88d8436e9e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190327
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins