sd: add documents for visually testing text fitting / scaling
Made this when changing the text scaling algorithm to have a
reference if we are regressing or progressing compared to MSO.
These documents are useful when further changing the text scaling
algorithm. Note that the rendering was not perfectly matching MSO,
but it was a progress. Making the algorithm be perfectly inline
with the MSO rendering is not a simple task.

Change-Id: I59a17d197724fd84603d8f2928c2a26ee2069a68
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190328
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins