BitmapEx->Bitmap in MetaBmpExScaleAction
now that Bitmap supports transparency

Disable part of the redaction test. I have
verified that the visual result of auto-image redaction
is the same before and after this change, but we
seem to end up with more rectangles than before.

Probably something weird inside the EMF/WMF machinery.
(see the fixMetaFile function).

Change-Id: I85f308ae66ddf7592e41fde1ab9e49f9c882146e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190338
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>