tdf#143182: fix string handling in SbxValue::Format
When a value is convertible to number, check if the format string
is for numbers of text (in SbxValue::Format). If the value is not
convertible to number, handle it according to the format string in
printfmtstr.

Change-Id: I6e11aeffd8dd8581ebc868762944e9d6d9393514
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190348
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit a3597edb52b671e6095f96c8705cd458a50d8799)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190386
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>