tdf#168145 Base must come out of the insert row
If you use ResultSets with the XRowLocate interface, Base moves to the insertion row in the WrappedResultSet.cxx file, line 73.

And it never returns. Normally, this shouldn't happen, after the insert is done you must go away from the insertion row to continue.
At least that's what it seems to me?

The same scenario happens in the file BookmarkSet.cxx line 78 and so this has also been corrected.

Change-Id: I169fd0a32441922025b3cc5531b32166b4c54be0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190349
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins