BitmapEx->Bitmap in FramePainter
now that Bitmap supports transparency

This flushes out an interesting bug. FramePainter::AdaptColor()
will set maCenter.maBitmap to empty. On a Bitmap that means
that the bitmap becomes 0x0 pixels. On a BitmapEx that means
that it has no data, but it thinks it still has the
same size.
Which in turns triggers a infinite loop in
OffsetBitmap::PaintCenter, so add a check for empty
there.

Change-Id: Iabe383818da9f38578e40ccfc452501a6c7da50d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190387
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>