a11y: Set a11y name for chart color palettes
    commit a26c93a6305bed437ab505713950858893db519d
    Date:   Sun Aug 10 22:33:37 2025 +0200

        chart color palettes: switch from drawingarea to IconView

had switched to using weld::IconView for the color
palettes.

Other than for the vcl IconView implementation, the gtk3
one does not use an item text for the accessible name
if no a11y name is explicitly set (even if an item text
were set, which is currently not the case for gtk3),
resulting in the Orca screen reader only announcing "icon"
when an item receives focus.

Explicitly set the accessible name for each item to the
same as the text set for the VCL variant to ensure an
a11y name is always set.

This now makes Orca e.g. announce "Palette 1, icon" when
that palette/item receives focus when using the gtk3
VCL plugin.

Sample steps to see the corresponding dialog page:

* start Writer
* click the "Insert Chart" toolbar item to insert
  a chart
* double-click on a bar in the chart
* switch to the "Color Palette" tab page

Change-Id: I641afe5b54055e91048fec111751f0ba3dafbea0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190368
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>