a11y: Drop manual a11y relations for chart color palettes
    commit a26c93a6305bed437ab505713950858893db519d
    Date:   Sun Aug 10 22:33:37 2025 +0200

        chart color palettes: switch from drawingarea to IconView

had switched to using weld::IconView for the color
palettes and introduced explicitly setting the label-for
and labelled-by accessible relations for the GtkFrame's
internal label child and the GtkIconView.

However, since the text of the GtkFrame's label is already
used as the accessible name for the frame and that one is the
parent "panel" of the IconView, screen readers already announce
the accessible name of the parent when the IconView receives
focus.

For gtk3, this would result in the label text/accessible name
getting announced twice now, e.g. "Colorful Colorful, list with
N items".
Similarly, NVDA on Windows would say "Colorful option pane, Colorful
list".

Stop explicitly setting the a11y relations to have it announced
only once.

An alternative solution could be to stop using a GtkFrame and
using a simple GtkLabel, then setting the GtkIconView as its
mnemonic widget.

Sample steps to see the corresponding dialog page:

* start Writer
* click the "Insert Chart" toolbar item to insert
  a chart
* double-click on a bar in the chart
* switch to the "Color Palette" tab page

Change-Id: If3a2beac193f05e13cc0fb1b30d6d0486c8433e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190369
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>