Unittests for pythonmaker
This patch adds the unit and integration test suite for the pythonmaker
tool.

The test is orchestrated by a Python script and uses a "golden file"
comparison method. It runs `pythonmaker` on a comprehensive test IDL
and verifies that the generated `.pyi` output exactly matches a set of
pre-validated reference stubs.

The test is integrated into the build system and runs with 'make check'.

Change-Id: If91814767779812740c543476307600f61c8a656
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188456
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>