tdf#168181: process field backslash escaping properly
For some reason, import code for HYPERLINK doesn't use vArguments from
the top of DomainMapper_Impl::CloseFieldCommand, and instead, it calls
pContext->GetCommandParts. But that gives escaped strings that must be
unescaped separately, according to the spec. There was a code handling
specifically file:/// URLs,  obviously to address the problem; but the
processing must be universal, both at import and export.

Change-Id: I4d91f1c14dd55077e73dca0141422ba4ebda83d3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190388
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>