tdf#168196 sw: copy header properties to all (first, even, odd)
Although a page style can have different CONTENTS
on left, right, and first pages,
the other header properties are the same/duplicated.

This patch tries to set up the plumbing,
and implements it for both the Borders and Background context dialog
as well as for the general Page Style dialog.

Unfortunately, our current implementation
does not use an actual shared format,
so the properties must be duplicated 4 times,
and yet sadly I couldn't find any existing mechanism
that attempted to keep them in sync.

Doing it properly doesn't sound like copy/paste,
so not something that I would attempt.

I don't see any interoperability reasons in DOC/DOCX
that would force the properties to be kept separate.
Certainly for ODF, it appears that only Master is exported
(in terms of the header and footer properties).

tdf#144448 was a nice duplicate report that pointed out the need
for FillHdFt to be used in order for Height/Spacing
to also be affected.
Otherwise SID_ATTR_PAGE_SIZE and SID_ATTR_PAGE_DYNAMIC
were just ignored as !GetRanges().doesContainWhich(nWhich)

Change-Id: I94f8cdf8bc6006dbea80941a2d6756e162f82411
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190420
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>