Fix tab item drawing failures on macOS Tahoe
Starting in macOS Tahoe, there are two issues that affect
drawing of tab items:

1. When compiled and run on macOS Tahoe, the tab item will
   fail to draw if the NSSegmentedControl is not a subview
   of a window whether the window is visible or not.
2. The selected tab item has the same bright background
   color as the default push button when the tab item is
   in the key window. In previous versions of macOS, the
   same background color is used for the selected tab
   item whether the tab item is enabled or not.

Both issues are fixed by adding the NSSegmentedControl
as a subview of a window. Note that the window must be
a key window to draw the selected tab item with the new
bright background color.

Also, macOS Tahoe has slightly larger tab items than previous
version so adjust tab item drawing bounds when built and run
on macOS Tahoe.

Change-Id: I56fa352d7a18f748e278f03c0e9c9bb483452466
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190438
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>