tdf#168172 sw markdown export: handle images
Load the bugdoc, save back to markdown, the image is lost.

Markdown doesn't seem to have a way to embed images, also these images
are not anchored.

So focus on the image type that is created by the markdown import:
linked, inline images.

This also required extending OutMarkdown_SwTextNode() a bit, because it
makes sense to not emit e.g. empty bold or italic hints, but e.g. inline
images are hints with no end, but we still want to produce output for
them.

Change-Id: I6529606ff4dd677815da4668f31c079184e18fa9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190475
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins