tdf#167564 sw markdown export: handle tables
Export the bugdoc to markdown, the content of the table is lost.

Common mark has no table support, but there is a spec at
<https://github.github.com/gfm/#tables-extension->, which seems to be
what md4c imports, so export that.

To make sure we export the table content exactly once, don't touch
m_pCurrentPam in OutMarkdown_SwTableNode(), just record the info we need
about the table, then let the iteration loop in SwMDWriter::Out_SwDoc()
process the table content.

This is just minimal support to make sure the table content is in the
output, table alignment is not implemented yet.

Change-Id: I9902fafc1ab5b8b7bc2472b026ab6df490f6407f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190514
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>