chart color palettes: No longer show palette names
    commit a26c93a6305bed437ab505713950858893db519d
    Date:   Sun Aug 10 22:33:37 2025 +0200

        chart color palettes: switch from drawingarea to IconView

had switched to using weld::IconView for the color
palettes and added setting a text, "Palette <number>"
for each of the palettes. The text was only displayed
for the VCL/non-gtk implementation, however.
Showing the text for gtk3 as well could be implemented
by adding a corresponding model column (see implementation
in earlier change set 1 of this Gerrit change, [1])

As Marco Cecchetti mentions in [2]:

> Anyway, I think more appropriate to remove the name
> in both cases since it's not descriptive and especially
> in the sidebar case it ends up just wasting space.
> Maybe you could set only for the a11y description.

Previous commit

    Change-Id: I641afe5b54055e91048fec111751f0ba3dafbea0
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Aug 29 13:43:40 2025 +0200

        a11y: Set a11y name for chart color palettes

implemented explicitly setting the a11y name to the
string that was previously set as the display text,
so the VCL fallback to use the display text for the
a11y name is no longer needed.

Therefore, drop the displayed text as suggested.

Sample steps to see the corresponding dialog page:

* start Writer
* click the "Insert Chart" toolbar item to insert
  a chart
* double-click on a bar in the chart
* switch to the "Color Palette" tab page

[1] https://gerrit.libreoffice.org/c/core/+/190367/1
[2] https://gerrit.libreoffice.org/c/core/+/190367/comments/d4b582df_0cc820cc

Change-Id: I5242219df6049ce8b6e99148d30cc9ea95485658
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190367
Tested-by: Jenkins
Reviewed-by: Marco Cecchetti <marco.cecchetti@collabora.com>