tdf#121253 Remove the default menu new document shortcut keys on MacOS
The default menu on MacOS is displayed when there are no windows or on a
dialog box with no parent. Previously there were hardcoded shortcut keys
to create new documents. As reported in tdf#121253, one of these is ⌘A
which ends up being inconvenient because that is commonly used to select
all the text in a text box. This had the effect that if you tried to use
that shortcut in some dialogs then the dialog would seem to disappear
and be replaced by the new database wizard.

The hardcoded shortcuts were added in 6efdd1444810 which is a patch from
January 2025. Before that the shortcuts were generated based on the menu
item title. ⌘A ends up being used for the database because ⌘D is already
used for a new Draw document and it just picked the next letter along.
These shortcuts were added in 2008 with ba57ba32e8054ec.

This patch removes the shortcuts for the default menu in both the copied
version of the start center menu and the original default menu generated
in code. It seems strange to have shortcuts for these only in the
specific case of there being no window or when a few select dialog boxes
are focused. If they were generally useful to people then surely it
would make sense to add them to the start center menu as well. For
people using non-English language packs the patch from January would
have effectively changed the shortcut keys. If nobody has complained
then it probably means they aren’t used.

Change-Id: I6065c3daa4e288caa98d73a3f286935dc5bc07ee
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190522
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
(cherry picked from commit 9d2da58ec93db75d2fcfce3dc001d9db0851e3c6)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190567
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>