tdf#167761 sw format redline, char style+direct: add ODT import
Open the bugdoc, it contains a format redline that changes the character
style from Strong to Quote and the font size from 24 to 36. Revert the
redline, the char style goes back to Strong, but the font size is 12,
not 24.

The problem is that only the char style was imported in
XMLRedlineImportHelper::InsertIntoDocument(), because direct formatting
is only handled when we have no char style.

Fix the problem by creating an empty item set earlier, similar to how
lcl_SetRedline() does it in the UI code, then both the char style and
direct format can go into that item set.

The format redline's ODT/DOCX filter is now in a reasonable shape with
this.

Change-Id: I28e2a43e218a319045649612e7492821e72649b5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190551
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit bbebb3908ff9a9c384a475737ead537906517387)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190556
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>