tdf#108911 Toggle input modes in Cell Edit handler, not input handler
The F2 key by default maps to “Cell Edit” which sets the input mode to
SC_INPUT_TABLE to start editing the cell. If you press F2 again, the key
event handling code in the input handler had a hardcoded handler for F2
which switched to SC_INPUT_TYPE mode when the mode is already
TABLE. This meant that you could press F2 to toggle between the two
modes. This is useful if you are editing an equation because sometimes
you want the cursor keys to move the text input cursor and sometimes you
want it to move the cell cursor to help type a cell reference.

This feature relies on the fact that the input handler code gets to see
the key event before the accelerator handler code. With the GTK backend
this works fine but on MacOS the accelerator code ends up eating the key
event before the input handler gets to see it. This meant that pressing
F2 always set TABLE mode instead of toggling. There’s also the slight
weirdness on all platforms that if you reassign “Cell Edit” to a
different shortcut key then it no longer works like a toggle and instead
you always have to press F2 to get to TYPE mode once you’re in TABLE
mode. It also means that if you tried to reassign F2 to a different
command then it wouldn’t work if you are editing a cell.

This patch changes the behaviour so that the toggling is done directly
in the handler for the “Cell Edit” command. The hardcoded handler for F2
in the input handler is removed. That means that the toggling will work
regardless of what key the command is assigned to and the input handler
doesn’t need to see the key before the accelerator key handler.

Change-Id: I79aa6fa515f7a720335f3535146c95eae874ea0a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190557
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>